fix number to bigint type error
strangely this only manifests in the github action. i would like to
revert this if possible but it probably doesn't matter anyways since the
seed is useless for our purposes.

+++ wasm/embind.cpp
@@ -102,7 +102,7 @@ EMSCRIPTEN_BINDINGS(tetris) {
     ;
 
   class_<Tetris>("Tetris")
-    .constructor<uint64_t>()
+    .constructor<unsigned int>()
     //.property("x", &MyClass::x)
     //.property("x", &MyClass::getX, &MyClass::setX)
     //.property("x_readonly", &MyClass::getX)
